Repo
- Updated `dub.json` so that when we are building in the default mode
(i.e.`debug`) only then is `gogga` included and the debugging versions
enabled. - This mode is present when running unittests and would make
all the `version)(unittest)` imports of `gogga` work. In the no-unit
test build case it would include `gogga` but of course the use cases of
it via `version(unittest)` would not be active